# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/material-submissions/{submissionId}": {
      "get": {
        "operationId": "MaterialsController_submissionDetail",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/submissions": {
      "post": {
        "operationId": "MaterialsController_createSubmission",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterialSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/files": {
      "post": {
        "operationId": "MaterialsController_addSubmissionFile",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadMaterialVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/files/{fileId}/remove": {
      "post": {
        "operationId": "MaterialsController_removeSubmissionFile",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveMaterialSubmissionFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/submit": {
      "post": {
        "operationId": "MaterialsController_submitSubmission",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/withdraw": {
      "post": {
        "operationId": "MaterialsController_withdrawSubmission",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawMaterialSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
...
    "/api/v1/material-submission-files/{fileId}/download": {
      "get": {
        "operationId": "MaterialsController_downloadSubmissionFile",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
...
    "/api/v1/applications": {
      "get": {
        "operationId": "ApplicationsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "HK_DIRECT",
                "JUPAS"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PLANNING",
                "CONFIRMED",
                "MATERIAL_PREPARATION",
                "PENDING_SUBMISSION",
                "SUBMITTED",
                "WAITING_RESULT",
                "SUPPLEMENT",
                "INTERVIEW",
                "OFFER",
                "REJECTED",
                "ENROLLED",
                "WITHDRAWN"
              ]
            }
          },
          {
            "name": "overdueOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "applications"
        ]
      },
      "post": {
        "operationId": "ApplicationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/portal/me/summary": {
      "get": {
        "operationId": "PortalController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/profile": {
      "get": {
        "operationId": "PortalController_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      },
      "post": {
        "operationId": "PortalController_submitProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPortalProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/materials": {
      "get": {
        "operationId": "PortalController_materials",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/materials/{materialId}/upload": {
      "post": {
        "operationId": "PortalController_upload",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalUploadMaterialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/material-versions/{versionId}/download": {
      "get": {
        "operationId": "PortalController_download",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/materials/{materialId}/submissions": {
      "post": {
        "operationId": "PortalController_createMaterialSubmission",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterialSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/material-submissions/{submissionId}": {
      "get": {
        "operationId": "PortalController_materialSubmissionDetail",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/material-submissions/{submissionId}/files": {
      "post": {
        "operationId": "PortalController_addMaterialSubmissionFile",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalUploadMaterialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/material-submissions/{submissionId}/files/{fileId}/remove": {
      "post": {
        "operationId": "PortalController_removeMaterialSubmissionFile",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveMaterialSubmissionFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/material-submissions/{submissionId}/submit": {
      "post": {
        "operationId": "PortalController_submitMaterialSubmission",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/material-submissions/{submissionId}/withdraw": {
      "post": {
        "operationId": "PortalController_withdrawMaterialSubmission",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawMaterialSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/materials/{materialId}/not-applicable-requests": {
      "post": {
        "operationId": "PortalController_requestMaterialNotApplicable",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestMaterialNotApplicableDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/material-submission-files/{fileId}/download": {
      "get": {
        "operationId": "PortalController_downloadMaterialSubmissionFile",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/progress": {
      "get": {
        "operationId": "PortalController_progress",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/applications": {
      "get": {
        "operationId": "PortalController_applications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/confirmations": {
      "get": {
        "operationId": "PortalController_confirmations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
    "/api/v1/portal/me/confirmations/{confirmationId}": {
      "post": {
        "operationId": "PortalController_respondConfirmation",
        "parameters": [
          {
            "name": "confirmationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "student-portal"
        ]
      }
    },
...
  },
  "info": {
    "title": "DSE升学服务CRM API",
    "description": "S1认证、权限、账号、审计与任务监督闭环接口",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dse_session"
      }
    },
    "schemas": {
...
      "UploadMaterialVersionDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "maxLength": 100
          },
          "contentBase64": {
            "type": "string",
            "description": "Base64编码文件内容，解码后最大50MB"
          },
          "replacesFileId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "补正时被替换的问题文件"
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "contentBase64"
        ]
      },
...
      "CreateMaterialSubmissionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "object",
            "maxLength": 1000,
            "nullable": true,
            "description": "管家代传时必填"
          }
        }
      },
      "RemoveMaterialSubmissionFileDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
      "WithdrawMaterialSubmissionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "PortalUploadMaterialDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "maxLength": 100
          },
          "contentBase64": {
            "type": "string"
          },
          "replacesFileId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "contentBase64"
        ]
      },
      "RequestMaterialNotApplicableDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "ApiEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "nullable": true
          },
          "error": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ],
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}